[lib] dnsmasq: add a test before installing packages

+++ src/vm5k/services/dnsmasq.py
@@ -107,15 +107,22 @@ def dnsmasq_server(server, clients=None, vms=None, dhcp=True):
 
     """
     logger.debug('Installing and configuring a DNS/DHCP server on %s', server)
-    cmd ='export DEBIAN_MASTER=noninteractive ; apt-get update ; apt-get -y purge dnsmasq-base ; '+\
-         'apt-get install -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confnew" '+\
-         '-y dnsmasq; echo 1 > /proc/sys/net/ipv4/ip_forward '
-    SshProcess(cmd, server).run()
+
+    test_running = Process('nmap ' + server + ' -p 53 | grep domain')
+    test_running.shell = True
+    test_running.run()
+    if 'domain' in test_running.stdout:
+        logger.info('DNS server already running, updating configuration')
+    else:
+        cmd ='export DEBIAN_MASTER=noninteractive ; apt-get update ; apt-get -y purge dnsmasq-base ; '+\
+             'apt-get install -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confnew" '+\
+             '-y dnsmasq; echo 1 > /proc/sys/net/ipv4/ip_forward '
+        SshProcess(cmd, server).run()
 
     vms_lists(vms, server)
     if clients:
         resolv_conf(server, clients)
-    if dhcp is not None:
+    if dhcp:
         sysctl_conf(server, vms)
         dhcp_conf(server, vms)
